+++ MqttLib/IMqttSubscriber.cs
@@ -12,14 +12,14 @@ namespace MqttLib
     /// </summary>
     /// <param name="subscriptions">Array of subscription objects</param>
     /// <returns>Message ID</returns>
-    int Subscribe(Subscription[] subscriptions);
+    ulong Subscribe(Subscription[] subscriptions);
 
     /// <summary>
     /// Subscribe to a single topic
     /// </summary>
     /// <param name="subscription">A Subscription object</param>
     /// <returns>Message ID</returns>
-    int Subscribe(Subscription subscription);
+    ulong Subscribe(Subscription subscription);
 
     /// <summary>
     /// Subscribe to a single topic
@@ -27,14 +27,14 @@ namespace MqttLib
     /// <param name="topic">Name of topic to subscribe to</param>
     /// <param name="qos">QoS</param>
     /// <returns>Message ID</returns>
-    int Subscribe(string topic, QoS qos);
+    ulong Subscribe(string topic, QoS qos);
 
     /// <summary>
     /// Unsubscribe from a topic
     /// </summary>
     /// <param name="topics">Topic Name</param>
     /// <returns>Message ID</returns>
-    int Unsubscribe(string[] topics);
+    ulong Unsubscribe(string[] topics);
 
     /// <summary>
     /// Fired when the Topic the MQTT client is subscribed to receives a message
